types

+++ src/main/java/com/invoiced/entity/Subscription.java
@@ -23,7 +23,7 @@ public final class Subscription extends AbstractEntity<Subscription> {
 
   @JsonInclude(JsonInclude.Include.NON_DEFAULT)
   @JsonProperty("amount")
-  public Long amount;
+  public Double amount;
 
   @JsonInclude(JsonInclude.Include.NON_EMPTY)
   @JsonProperty("plan")
